# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cliente API
  version: 1.0.0
  description: >-
    API REST para cadastro, autenticação e gerenciamento de clientes. O token de
    acesso segue o formato JSON Web Token (JWT) e é enviado no esquema Bearer do
    cabeçalho Authorization. Os papéis `customer` e `admin` definem quais
    clientes cada pessoa pode consultar ou modificar.
servers:
  - url: http://127.0.0.1:8080
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      tags:
        - Autenticação
      summary: Cadastrar cliente e retornar tokens de sessão
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "201":
          description: Cliente cadastrado e autenticado
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControlNoStore"
            Pragma:
              $ref: "#/components/headers/PragmaNoCache"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
        "408":
          $ref: "#/components/responses/RequestCanceled"
        "504":
          $ref: "#/components/responses/Timeout"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/login:
    post:
      tags:
        - Autenticação
      summary: Autenticar cliente e retornar tokens de sessão
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Cliente autenticado
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControlNoStore"
            Pragma:
              $ref: "#/components/headers/PragmaNoCache"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "429":
          $ref: "#/components/responses/RateLimited"
        "408":
          $ref: "#/components/responses/RequestCanceled"
        "504":
          $ref: "#/components/responses/Timeout"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/refresh:
    post:
      tags:
        - Autenticação
      summary: Renovar o par de tokens com rotação do refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Tokens renovados; o refresh token anterior foi revogado
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControlNoStore"
            Pragma:
              $ref: "#/components/headers/PragmaNoCache"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "429":
          $ref: "#/components/responses/RateLimited"
        "408":
          $ref: "#/components/responses/RequestCanceled"
        "504":
          $ref: "#/components/responses/Timeout"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/logout:
    post:
      tags:
        - Autenticação
      summary: Revogar a família do refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "204":
          description: Sessão revogada ou já inexistente
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControlNoStore"
            Pragma:
              $ref: "#/components/headers/PragmaNoCache"
        "400":
          $ref: "#/components/responses/BadRequest"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "429":
          $ref: "#/components/responses/RateLimited"
        "408":
          $ref: "#/components/responses/RequestCanceled"
        "504":
          $ref: "#/components/responses/Timeout"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /cliente:
    get:
      tags:
        - Clientes
      summary: Listar clientes como administrador
      responses:
        "200":
          description: Lista de clientes
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControlNoStore"
            Pragma:
              $ref: "#/components/headers/PragmaNoCache"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "408":
          $ref: "#/components/responses/RequestCanceled"
        "504":
          $ref: "#/components/responses/Timeout"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - Clientes
      summary: Criar cliente como administrador
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "201":
          description: Cliente criado
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControlNoStore"
            Pragma:
              $ref: "#/components/headers/PragmaNoCache"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "409":
          $ref: "#/components/responses/Conflict"
        "408":
          $ref: "#/components/responses/RequestCanceled"
        "504":
          $ref: "#/components/responses/Timeout"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /cliente/{id}:
    get:
      tags:
        - Clientes
      summary: Buscar o próprio cliente ou qualquer cliente como administrador
      parameters:
        - $ref: "#/components/parameters/CustomerID"
      responses:
        "200":
          description: Cliente encontrado
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControlNoStore"
            Pragma:
              $ref: "#/components/headers/PragmaNoCache"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "408":
          $ref: "#/components/responses/RequestCanceled"
        "504":
          $ref: "#/components/responses/Timeout"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      tags:
        - Clientes
      summary: Atualizar o próprio cliente ou qualquer cliente como administrador
      parameters:
        - $ref: "#/components/parameters/CustomerID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerRequest"
      responses:
        "200":
          description: Cliente atualizado
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControlNoStore"
            Pragma:
              $ref: "#/components/headers/PragmaNoCache"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "408":
          $ref: "#/components/responses/RequestCanceled"
        "504":
          $ref: "#/components/responses/Timeout"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - Clientes
      summary: Remover o próprio cliente ou qualquer cliente como administrador
      parameters:
        - $ref: "#/components/parameters/CustomerID"
      responses:
        "204":
          description: Cliente removido
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControlNoStore"
            Pragma:
              $ref: "#/components/headers/PragmaNoCache"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "408":
          $ref: "#/components/responses/RequestCanceled"
        "504":
          $ref: "#/components/responses/Timeout"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    CacheControlNoStore:
      description: Impede o armazenamento de respostas autenticadas ou com credenciais.
      schema:
        type: string
        enum:
          - no-store
    PragmaNoCache:
      description: Reforça a instrução de não armazenar a resposta em cache.
      schema:
        type: string
        enum:
          - no-cache
  parameters:
    CustomerID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: 019f41d1-d56a-7d2e-bf8f-240825a21106
  responses:
    BadRequest:
      description: Requisição inválida
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControlNoStore"
        Pragma:
          $ref: "#/components/headers/PragmaNoCache"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Token ausente, inválido ou credenciais inválidas
      headers:
        WWW-Authenticate:
          description: Desafio Bearer para rotas protegidas; pode não existir em erros de login ou refresh token.
          schema:
            type: string
        Cache-Control:
          $ref: "#/components/headers/CacheControlNoStore"
        Pragma:
          $ref: "#/components/headers/PragmaNoCache"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: A identidade autenticada não possui permissão para a operação
      headers:
        WWW-Authenticate:
          description: Desafio Bearer com insufficient_scope
          schema:
            type: string
        Cache-Control:
          $ref: "#/components/headers/CacheControlNoStore"
        Pragma:
          $ref: "#/components/headers/PragmaNoCache"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Cliente não encontrado
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControlNoStore"
        Pragma:
          $ref: "#/components/headers/PragmaNoCache"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Email já existe
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControlNoStore"
        Pragma:
          $ref: "#/components/headers/PragmaNoCache"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    PayloadTooLarge:
      description: Corpo JSON maior que 64 KiB
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControlNoStore"
        Pragma:
          $ref: "#/components/headers/PragmaNoCache"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnsupportedMediaType:
      description: Content-Type diferente de application/json
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControlNoStore"
        Pragma:
          $ref: "#/components/headers/PragmaNoCache"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Limite de requisições por IP excedido
      headers:
        Retry-After:
          description: Segundos mínimos antes de uma nova tentativa
          schema:
            type: integer
            minimum: 1
        Cache-Control:
          $ref: "#/components/headers/CacheControlNoStore"
        Pragma:
          $ref: "#/components/headers/PragmaNoCache"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Timeout:
      description: A operação ultrapassou o limite de 10 segundos
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControlNoStore"
        Pragma:
          $ref: "#/components/headers/PragmaNoCache"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RequestCanceled:
      description: A requisição foi cancelada antes de sua conclusão
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControlNoStore"
        Pragma:
          $ref: "#/components/headers/PragmaNoCache"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Erro interno inesperado, sem exposição de detalhes sensíveis
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControlNoStore"
        Pragma:
          $ref: "#/components/headers/PragmaNoCache"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: validation_error
        message:
          type: string
          example: email deve ser válido
    CreateCustomerRequest:
      type: object
      additionalProperties: false
      required:
        - name
        - email
        - phone
        - password
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 255
          pattern: '^\S(.*\S)?$'
          description: O contrato exige o nome sem espaços nas extremidades; o servidor também os remove defensivamente antes de validar.
          example: Admin Aula
        email:
          type: string
          format: email
          maxLength: 255
          pattern: '^[a-z0-9._%+\-]+@[a-z0-9.\-]+\.[a-z]{2,}$'
          description: O contrato exige a forma canônica em minúsculas e sem espaços nas extremidades; o servidor também normaliza defensivamente.
          example: admin@example.com
        phone:
          type: string
          minLength: 10
          maxLength: 30
          pattern: '^\+?[0-9(][0-9 ().-]*[0-9)]$'
          description: O contrato exige o valor sem espaços nas extremidades e com 10 a 15 dígitos; espaços internos e os separadores (), ponto e hífen são aceitos. O servidor remove espaços externos defensivamente.
          example: (11) 98888-7777
        password:
          type: string
          format: password
          minLength: 8
          pattern: '^.*\S.*$'
          description: Deve conter ao menos um caractere que não seja espaço e ter no máximo 72 bytes quando codificada em UTF-8, limite do bcrypt.
          example: senha-forte-123
    UpdateCustomerRequest:
      type: object
      additionalProperties: false
      required:
        - name
        - email
        - phone
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 255
          pattern: '^\S(.*\S)?$'
          description: O contrato exige o nome sem espaços nas extremidades; o servidor também os remove defensivamente antes de validar.
          example: Cliente Atualizado
        email:
          type: string
          format: email
          maxLength: 255
          pattern: '^[a-z0-9._%+\-]+@[a-z0-9.\-]+\.[a-z]{2,}$'
          description: O contrato exige a forma canônica em minúsculas e sem espaços nas extremidades; o servidor também normaliza defensivamente.
          example: cliente@example.com
        phone:
          type: string
          minLength: 10
          maxLength: 30
          pattern: '^\+?[0-9(][0-9 ().-]*[0-9)]$'
          description: O contrato exige o valor sem espaços nas extremidades e com 10 a 15 dígitos; espaços internos e os separadores (), ponto e hífen são aceitos. O servidor remove espaços externos defensivamente.
          example: (21) 98888-1111
    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          pattern: '^[a-z0-9._%+\-]+@[a-z0-9.\-]+\.[a-z]{2,}$'
          description: O contrato exige a forma canônica em minúsculas e sem espaços nas extremidades; o servidor também normaliza defensivamente antes de autenticar.
          example: admin@example.com
        password:
          type: string
          format: password
          minLength: 8
          pattern: '^.*\S.*$'
          description: Deve conter ao menos um caractere que não seja espaço e ter no máximo 72 bytes quando codificada em UTF-8, limite do bcrypt.
          example: senha-forte-123
    RefreshTokenRequest:
      type: object
      additionalProperties: false
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          minLength: 46
          maxLength: 46
          pattern: '^rt_[A-Za-z0-9_-]{43}$'
          description: Refresh token opaco retornado no login, cadastro ou renovação.
          example: rt_6jvWz5vPqm9V4gxKbtRk0K5ZgW6dN5PqL3sB8C1xY2A
    Customer:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - email
        - phone
        - role
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        role:
          type: string
          enum:
            - customer
            - admin
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AuthResponse:
      type: object
      additionalProperties: false
      required:
        - access_token
        - token_type
        - expires_at
        - refresh_token
        - refresh_expires_at
        - session_expires_at
        - customer
      properties:
        access_token:
          type: string
          description: JWT assinado.
        token_type:
          type: string
          enum:
            - Bearer
          example: Bearer
        expires_at:
          type: string
          format: date-time
        refresh_token:
          type: string
          minLength: 46
          maxLength: 46
          pattern: '^rt_[A-Za-z0-9_-]{43}$'
          description: Token opaco de uso único; armazene-o de forma segura e substitua-o a cada renovação.
        refresh_expires_at:
          type: string
          format: date-time
          description: >-
            Limite de inatividade atual do refresh token, renovado a cada
            rotação sem ultrapassar a sessão.
        session_expires_at:
          type: string
          format: date-time
          description: Limite absoluto da família; não é prorrogado pelas rotações.
        customer:
          $ref: "#/components/schemas/Customer"